+++ src/main.rs
@@ -1,9 +1,4 @@
-pub mod ast;
-pub mod automata;
-pub mod formalism;
+pub mod ndfa;
 
 fn main() {
-    let regex = "(ab|b)*a|a(a|b)b*";
-    let (linearized_symbols, ast_root) = ast::parse(&regex);
-    automata::NDFA::new(&linearized_symbols, &ast_root);
 }
